Moving display code to _head
Wiiceiver_master is a display-aware Wiiceiver base.  Sadly, it’s up
against memory limits.  It sends a buffer (struct) out the i2c hole.

Wiiceiver_head is a totally different codebase, designed to drive / act
as a frame buffer for an Adafruit display based on SSD1306.
Wiiceiver_head is intended to run on a separate Arduino (possibly pro
mini), SPI-attached to the display.