catching up; Board + Chuck (sorta) work
significant catchup.  Chuck has a compile-time way to act like a
nunchuck or a “fake” chuck.  Lots of work needed for cleanup, but it’s
there.

Board has a way to FORCE heli-mode … not very pretty, there’s not a
good way to un-force.  Need to work on this more.  otherwise, also
works.  Enabled for quad ESC now, need to test.